2024/5/7 後台功能和html

+++ team6/src/main/java/com/team6/delivery/model/Delivery.java
@@ -1,7 +1,5 @@
 package com.team6.delivery.model;
 
-import org.hibernate.annotations.ManyToAny;
-import org.springframework.data.annotation.Transient;
 import org.springframework.stereotype.Component;
 
 import jakarta.persistence.Column;
@@ -9,8 +7,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
 @Entity
@@ -26,7 +22,7 @@ public class Delivery {
 	private String orderid;
 	private String date;
 	private String address;
-	private byte status = 1;
+	private int status = 1;
 	
 
 
@@ -66,11 +62,11 @@ public class Delivery {
 		this.address = address;
 	}
 
-	public byte getStatus() {
+	public int getStatus() {
 		return status;
 	}
 
-	public void setStatus(byte status) {
+	public void setStatus(int status) {
 		this.status = status;
 	}
